+++ tests/pages/base_page.py
@@ -1,7 +1,5 @@
 from selenium.common.exceptions import StaleElementReferenceException, InvalidSelectorException
 from selenium.webdriver.support.wait import WebDriverWait
-import random
-import string
 
 
 def visibility_of_element(locator):
@@ -15,15 +13,6 @@ def visibility_of_element(locator):
             return False
 
     return _predicate
-
-
-def random_string(string_len=7):
-    sets = string.ascii_uppercase + string.digits + string.ascii_lowercase
-    return ''.join(random.choices(sets, k=string_len))
-
-
-def random_email():
-    return 'ex_test_' + random_string(string_len=12) + "@yopmail.com"
 
 
 class BasePage:
