+++ tests/pages/item_in_catalog.py
@@ -1,4 +1,3 @@
-from selenium.common.exceptions import InvalidSelectorException, StaleElementReferenceException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.remote.webdriver import WebDriver
 from selenium.webdriver.support.wait import WebDriverWait
