+++ tests/pages/main_page.py
@@ -20,20 +20,17 @@ class MainPage(BasePage):
         self.page_url = base_url + self.URL
 
     def slider_is_visible(self) -> bool:
-        element = self.driver.find_element(*self.SLIDER)
-        return element.is_displayed()
+        return self.driver.find_element(*self.SLIDER).is_displayed()
 
     def features_is_visible(self) -> bool:
         element = self.driver.find_element(*self.FEATURED)
         return element.is_displayed()
 
     def brand_swiper_is_visible(self) -> bool:
-        element = self.driver.find_element(*self.BRAND_SWIPER)
-        return element.is_displayed()
+        return self.driver.find_element(*self.BRAND_SWIPER).is_displayed()
 
     def slider_pagination_is_visible(self) -> bool:
-        element = self.driver.find_element(*self.SLIDER_PAGINATION)
-        return element.is_displayed()
+        return self.driver.find_element(*self.SLIDER_PAGINATION).is_displayed()
 
     def get_title(self):
         return self.driver.title
@@ -45,5 +42,4 @@ class MainPage(BasePage):
         self.get_element(self.STERLING_CURRENCY).click()
 
     def sterling_currency_is_visible(self) -> bool:
-        element = self.driver.find_element(*self.STERLING_CURRENCY_SELECTED)
-        return element.is_displayed()
+        return self.driver.find_element(*self.STERLING_CURRENCY_SELECTED).is_displayed()
